Move language selection to shared module
