+++ services/cms/package.json
@@ -31,6 +31,7 @@
     "@mui/lab": "^5.0.0-alpha.68",
     "@mui/material": "^5.6.4",
     "@mui/styles": "^5.6.2",
+    "@popperjs/core": "^2.11.5",
     "@wordpress/base-styles": "^4.4.0",
     "@wordpress/block-editor": "^9.0.0",
     "@wordpress/block-library": "^7.5.0",
@@ -59,6 +60,7 @@
     "react-hook-form": "^7.30.0",
     "react-hot-toast": "^2.2.0",
     "react-i18next": "^11.16.9",
+    "react-popper": "^2.3.0",
     "react-query": "^3.39.0",
     "react-spring": "^9.4.5",
     "recoil": "^0.7.2",
