+++ services/course-material/package.json
@@ -24,6 +24,7 @@
     "@mui/lab": "^5.0.0-alpha.68",
     "@mui/material": "^5.8.2",
     "@mui/styles": "^5.8.0",
+    "@popperjs/core": "^2.11.5",
     "axios": "^0.27.2",
     "citation-js": "^0.5.7",
     "date-fns": "^2.28.0",
@@ -43,6 +44,7 @@
     "react-i18next": "^11.17.0",
     "react-is": "^18.1.0",
     "react-medium-image-zoom": "^4.4.3",
+    "react-popper": "^2.3.0",
     "react-query": "^3.39.1",
     "react-spring": "^9.4.5",
     "react-tsparticles": "^2.0.6",
