+++ services/example-exercise/package.json
@@ -25,6 +25,7 @@
     "@mui/lab": "^5.0.0-alpha.68",
     "@mui/material": "^5.8.2",
     "@mui/styles": "^5.8.0",
+    "@popperjs/core": "^2.11.5",
     "date-fns": "^2.28.0",
     "express": "^4.18.1",
     "http-proxy-middleware": "^2.0.6",
@@ -38,6 +39,7 @@
     "react-hot-toast": "^2.2.0",
     "react-i18next": "^11.17.0",
     "react-is": "^18.1.0",
+    "react-popper": "^2.3.0",
     "react-spring": "^9.4.5",
     "svgo": "^2.8.0",
     "use-debounce": "^8.0.1",
