+++ services/quizzes/package.json
@@ -27,6 +27,7 @@
     "@mui/lab": "^5.0.0-alpha.68",
     "@mui/material": "^5.8.2",
     "@next/bundle-analyzer": "^12.1.6",
+    "@popperjs/core": "^2.11.5",
     "@types/commonmark": "^0.27.5",
     "axios": "^0.27.2",
     "commonmark": "^0.30.0",
@@ -45,6 +46,7 @@
     "react-hook-form": "^7.31.3",
     "react-hot-toast": "^2.2.0",
     "react-i18next": "^11.17.0",
+    "react-popper": "^2.3.0",
     "react-query": "^3.39.1",
     "react-redux": "^8.0.2",
     "react-spring": "^9.4.5",
