+++ shared-module/package.json
@@ -63,6 +63,7 @@
     "@mui/lab": "^5.0.0-alpha.68",
     "@mui/material": "^5.6.4",
     "@mui/styles": "^5.6.2",
+    "@popperjs/core": "^2.11.5",
     "axios": "^0.27.2",
     "date-fns": "^2.28.0",
     "i18next": "^21.8.0",
@@ -71,6 +72,7 @@
     "react-hook-form": "^7.30.0",
     "react-hot-toast": "^2.2.0",
     "react-i18next": "^11.16.9",
+    "react-popper": "^2.3.0",
     "react-query": "^3.39.0",
     "react-spring": "^9.4.5",
     "react-tsparticles": "^2.0.6",
