+++ shared-module/src/locales/en/shared-module.json
@@ -5,6 +5,7 @@
   "available": "Available",
   "breadcrumb": "Breadcrumb",
   "chapter-chapter-number": "Chapter {{number}}",
+  "choose-a-language": "Choose language",
   "close": "Close",
   "closed": "Closed",
   "courses": "Courses",
