+++ shared-module/src/locales/fi/shared-module.json
@@ -5,6 +5,7 @@
   "available": "Saatavilla",
   "breadcrumb": "Muropolku",
   "chapter-chapter-number": "Luku {{number}}",
+  "choose-a-language": "Valitse kieli",
   "close": "Sulje",
   "closed": "Suljettu",
   "courses": "Kurssit",
